# apps/api/app/resources/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "REST API with Vapor",
    "title": "NXGT Vapor",
    "version": "1.0.0"
  },
  "servers": [
    {
      "description": "Dev server",
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "List all users.",
        "operationId": "getUsers",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/size"
          },
          {
            "$ref": "#/components/parameters/paged"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/UserPage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "Create a new user.",
        "operationId": "createUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/groups": {
      "get": {
        "tags": [
          "groups-controller"
        ],
        "summary": "List all groups.",
        "operationId": "getGroups",
        "parameters": [
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/size"
          },
          {
            "$ref": "#/components/parameters/paged"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of groups.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/GroupPage"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "groups-controller"
        ],
        "summary": "Create a new group.",
        "operationId": "createGroup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the permission.",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "The name of the permission."
          },
          "description": {
            "type": "string",
            "description": "The description of the permission."
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the role.",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "description": "The name of the role."
          },
          "description": {
            "type": "string",
            "description": "The description of the role."
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "description": "An array of permission objects that represent the role's permissions."
          }
        },
        "required": [
          "name"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the user."
          },
          "username": {
            "type": "string",
            "description": "The user's username."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The user's email address."
          },
          "password": {
            "type": "string",
            "description": "The user's password."
          },
          "avatar": {
            "type": "string",
            "format": "binary",
            "description": "The user's avatar image, encoded in Base64."
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "description": "An array of role objects that represent the user's roles."
          }
        },
        "required": ["username", "email", "password"]
      },
      "UserRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "The user's username."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The user's email address."
          },
          "password": {
            "type": "string",
            "description": "The user's password."
          },
          "avatar": {
            "type": "string",
            "format": "binary",
            "description": "The user's avatar image."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The user's roles."
          }
        }
      },
...
      "UserPage": {
        "description": "User page",
        "properties": {
          "items": {
            "description": "User list",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "Group": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier of the group."
          },
          "name": {
            "type": "string",
            "description": "The name of the group."
          },
          "description": {
            "type": "string",
            "description": "A description of the group."
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "description": "An array of role objects that represent the group's roles."
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "description": "An array of user objects that represent the group's users."
          }
        },
        "required": ["name"]
      },
      "GroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the group."
          },
          "description": {
            "type": "string",
            "description": "The description of the group."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The group's roles."
          },
          "users": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The group's users."
          }
        },
        "required": [
          "name"
        ]
      },
...
      "GroupPage": {
        "description": "Group page",
        "properties": {
          "items": {
            "description": "Group list",
            "items": {
              "$ref": "#/components/schemas/Group"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "The current page number."
          },
          "size": {
            "type": "integer",
            "description": "The number of items per page."
          },
          "total": {
            "type": "integer",
            "description": "The total number of items."
          },
          "pageCount": {
            "type": "integer",
            "description": "The total number of pages."
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "pageCount"
        ]
      },
      "ErrorModel": {
        "description": "Error payload",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "The user is not authorized to access the requested resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The user does not have permission to access the requested resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unexpected error occurred on the server.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      }
    },
    "parameters": {
...
      "page": {
        "in": "query",
        "name": "page",
        "description": "The page number to return.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "size": {
        "in": "query",
        "name": "size",
        "description": "The number of items to return per page.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "paged": {
        "in": "query",
        "name": "paged",
        "description": "Whether to return the results as a paged response.",
        "schema": {
          "type": "boolean"
        }
      }
    },
...
  }
}